[Feature #166179115] Rename party view module to editParty view
