+++ UI/edit_party.html
@@ -101,7 +101,7 @@
   <script src="js/api/formValidation.js"></script>
   <script type="module" src="js/api/helpers.js"></script>
   <script type="module" src="js/api/view/notificationToast.js"></script>
-  <script type="module" src="js/api/view/party.js"></script>
+  <script type="module" src="js/api/view/editParty.js"></script>
   <script type="module" src="js/api/model/party.js"></script>
   <script type="module" src="js/api/controller/editParty.js"></script>
   <script type="module" src="js/api/logOut.js"></script>
